relocate til list and item components
