+++ app/til/[id]/page.tsx
@@ -1,12 +1,9 @@
-import { Suspense } from "react"
 import { Metadata } from "next"
-import Link from "next/link"
-import { MDXRemote } from "next-mdx-remote/rsc"
 
 import { getAllTILIds, getAllTILs } from "@/lib/github-api"
-import { cn, getOgImgUrl } from "@/lib/utils"
-import { badgeVariants } from "@/components/ui/badge"
-import { MDXTIL } from "@/components/mdx-components"
+import { getOgImgUrl } from "@/lib/utils"
+
+import { TILItem } from "../components"
 
 export async function generateStaticParams() {
   let ids = getAllTILIds()
@@ -68,55 +65,4 @@ export default async function TilItemPage({
   if (!til) return null
 
   return <TILItem til={til} />
-}
-
-export function TILItem({
-  til,
-  ...props
-}: React.ComponentProps<"div"> & {
-  til: {
-    id: string
-    title: string
-    body: string
-    bodyHTML: string
-    createdAt: string
-    labels: { nodes: { name: string }[] }
-  }
-}) {
-  return (
-    <div
-      key={til.id}
-      className="break-words border-b border-b-gray-200 pb-6"
-      {...props}
-    >
-      <Link className="no-underline" href={`/til/${til.id}`}>
-        <h2 id={til.id} className="text-md font-mono">
-          <div className="pb-4 text-sm font-normal">
-            {new Date(til.createdAt)?.toLocaleDateString("ko")}
-          </div>
-          {til.title}
-        </h2>
-      </Link>
-      <Suspense>
-        {/* @ts-expect-error */}
-        <MDXTIL source={til.body} />
-      </Suspense>
-      <div className="mb-2 flex gap-2 pt-5">
-        {til.labels.nodes.map((label: any) => (
-          <Link
-            className={cn(
-              badgeVariants({
-                variant: "outline",
-              }),
-              "no-underline"
-            )}
-            href={`/til/label/${label.name}`}
-            key={label.id}
-          >
-            {label.name}
-          </Link>
-        ))}
-      </div>
-    </div>
-  )
 }
